+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-07-13
+
+### Added
+
+- Added Tolerable Creepers integration!
+    - Adapted Mobs now have their own Creepie variants, spores, and spore barrels.
+
+### Changed
+
+- Reworked Festive, Support, and Rocket Creeper models/textures.
+- Many miscellaneous tweaks and fixes.
+
 ## [1.0.0] - 2026-07-06
 
 - Initial release.